Payara-1554 Addressing Andy's feedback

+++ appserver/payara-appserver-modules/microprofile/jwt-auth/src/main/java/fish/payara/microprofile/jwtauth/jwt/JwtTokenParser.java
@@ -141,7 +141,7 @@ public class JwtTokenParser {
 
     private boolean checkNotExpired(Map<String, JsonValue> presentedClaims) {
         int currentTime = (int) (System.currentTimeMillis() / 1000);
-        long expiredTime = ((JsonNumber) presentedClaims.get(exp.name())).intValue();
+        int expiredTime = ((JsonNumber) presentedClaims.get(exp.name())).intValue();
 
         return currentTime < expiredTime;
     }
